Move ATM SABR vol method to Hagan2002LognormalSABR objects.
