+++ pysabr/tests/hagan_2002_lognormal_sabr/test_interpolation.py
@@ -38,7 +38,8 @@ def test_lognormal_beta_05_smile():
         [10.0000, 2.5271, 3.0000, 0.0253, 0.5000, -0.2463, 0.2908])
     k = (k + s) / 100
     f = (f + s) / 100
-    vols_test = sabr.lognormal_vol(k, f, t, alpha, beta, rho, volvol) * 100
+    vols_test = [sabr.lognormal_vol(k_, f, t, alpha, beta, rho, volvol) * 100
+                 for k_ in k]
     vols_target = np.array([19.641923, 15.785344, 14.305103, 13.073869,
                             12.550007, 12.088721, 11.691661, 11.517660,
                             11.360133, 11.219058, 11.094293, 10.892464,
